cluster: refactors to accommodate pke/azure integration
