+++ cluster/nodelabels.go
@@ -16,6 +16,7 @@ package cluster
 
 import (
 	"context"
+	"regexp"
 	"strconv"
 	"strings"
