+++ pkg/k8sutil/k8sutil.go
@@ -267,6 +267,8 @@ func IsK8sErrorPermanent(err error) bool {
 		return false // Newly instantiated AKS cluster's ETCD is flaky
 	} else if strings.Contains(err.Error(), "connection refused") {
 		return false
+	} else if strings.Contains(err.Error(), " i/o timeout") {
+		return false
 	}
 
 	return true
